# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Kokak API - Philippine language word puzzle game
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Language {
  id: number;
  code: string;
  name: string;
  wordCount: number;
}

export interface Puzzle {
  id: number;
  filipinoWord: string;
  targetWord: string;
  wordLength: number;
  languageCode: string;
  languageName: string;
}

export interface PuzzleDetail {
  id: number;
  filipinoWord: string;
  targetWord: string;
  wordLength: number;
  languageCode: string;
  languageName: string;
  meaning: string;
  pronunciation: string;
  exampleSentence: string;
  /** @nullable */
  usageNotes?: string | null;
}

export type ContributionStatus = typeof ContributionStatus[keyof typeof ContributionStatus];


export const ContributionStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Contribution {
  id: number;
  word: string;
  languageCode: string;
  meaning: string;
  status: ContributionStatus;
  submittedAt: string;
  /** @nullable */
  aiValidationNote?: string | null;
}

export interface ContributionInput {
  /** @minLength 1 */
  word: string;
  /** @minLength 1 */
  languageCode: string;
  /** @minLength 1 */
  meaning: string;
  /** @nullable */
  filipinoEquivalent?: string | null;
}

export type GetRandomPuzzleParams = {
/**
 * The language code (e.g. ceb, ilo, hil)
 */
languageCode: string;
};