# lib/api-zod/src/generated/types/contribution.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Kokak API - Philippine language word puzzle game
 * OpenAPI spec version: 0.1.0
 */
import type { ContributionStatus } from './contributionStatus';

export interface Contribution {
  id: number;
  word: string;
  languageCode: string;
  meaning: string;
  status: ContributionStatus;
  submittedAt: Date;
  /** @nullable */
  aiValidationNote?: string | null;
}