# lib/api-zod/src/generated/types/contributionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Kokak API - Philippine language word puzzle game
 * OpenAPI spec version: 0.1.0
 */

export interface ContributionInput {
  /** @minLength 1 */
  word: string;
  /** @minLength 1 */
  languageCode: string;
  /** @minLength 1 */
  meaning: string;
  /** @nullable */
  filipinoEquivalent?: string | null;
}