# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Kokak API - Philippine language word puzzle game
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}