# lib/api-zod/src/generated/types/language.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Kokak API - Philippine language word puzzle game
 * OpenAPI spec version: 0.1.0
 */

export interface Language {
  id: number;
  code: string;
  name: string;
  wordCount: number;
}