# lib/api-zod/src/generated/types/puzzle.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Kokak API - Philippine language word puzzle game
 * OpenAPI spec version: 0.1.0
 */

export interface Puzzle {
  id: number;
  filipinoWord: string;
  targetWord: string;
  wordLength: number;
  languageCode: string;
  languageName: string;
}