# lib/api-zod/src/generated/types/puzzleDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Kokak API - Philippine language word puzzle game
 * OpenAPI spec version: 0.1.0
 */

export interface PuzzleDetail {
  id: number;
  filipinoWord: string;
  targetWord: string;
  wordLength: number;
  languageCode: string;
  languageName: string;
  meaning: string;
  pronunciation: string;
  exampleSentence: string;
  /** @nullable */
  usageNotes?: string | null;
}